ux_rework_add_home_search Add the search in the home tab
